+++ HISTORY.rst
@@ -7,11 +7,14 @@ History
 - Added computational grid checks before an upload (#429).
 - Added handling of the 'started_from' Simulation parameter (#556).
 - Updated threedi-api-client version to 4.1.7.
+- Added simulation name sanitization (#497).
+- Changed simulation results directory name (#530).
 
 
 3.10.2 (2024-06-05)
 -------------------
 
+- Add substance concentrations to Boundary Conditions page (#559).
 - Use column names i/o column orderings to read CSV uploaded files (#563).
 - Add substance concentrations to Laterals page (#553, #557).
 - Add substances page to the simulation wizard (#548, #554).